Code adjustments for pacman-4.1 API changes

+++ src/kalu/conf.h
@@ -58,7 +58,7 @@ typedef struct _pacman_config_t {
     char            *arch;
     int              checkspace;
     int              usesyslog;
-    int              usedelta;
+    double           usedelta;
     alpm_list_t     *ignorepkgs;
     alpm_list_t     *ignoregroups;
     alpm_list_t     *noupgrades;
